+++ pipeline/tick/ast.go
@@ -96,6 +96,8 @@ func (a *AST) Create(n pipeline.Node, parents []ast.Node) (ast.Node, error) {
 		return NewDelete(parents).Build(node)
 	case *pipeline.DerivativeNode:
 		return NewDerivative(parents).Build(node)
+	case *pipeline.Ec2AutoscaleNode:
+		return NewEc2Autoscale(parents).Build(node)
 	case *pipeline.EvalNode:
 		return NewEval(parents).Build(node)
 	case *pipeline.FlattenNode: